refactor: cargo clippy

+++ src/drivers/disk/driver.rs
@@ -97,7 +97,7 @@ pub fn read_from_partition(
 ) -> Result<Vec<u8>, DiskError> {
     let start_sector = start_sector + partition.1;
     assert!(
-        (start_sector + sector_count as u64) < partition.1 + partition.2,
+        (start_sector + sector_count) < partition.1 + partition.2,
         "Trying to read outside of partition"
     );
     read_from_disk(&partition.0, start_sector, sector_count)
